refactor(Count, Money): vo 패키지 생성 및 클래스 이동 - 값 표현용 객체와 엔티티를 구분하기 위해 vo 패키지 생성
